chore: updates the compiled assets based on the current source
I followed the instructions to build the compiled assets for this package when I was working on a
completely different change. I noticed that the version controlled compiled assets are relatively
out of sync with the compiled version of the src. Thus, this PR is just to sync those up. Note that
I made no changes to the src, so all of the compiled assets _should_ be equivalent. I did not check
all of them because there were so many changes.