chore: moves Fractal configuration to its own package as well.
Next, I moved all of the Fractal related files to the `fractal/` folder.
To facilitate a less redundant package structure, I added a
`package.json` and made this dependent on the `stjudecloud/@assets`
package.

Note that I replaced our custom build system with Parcel v2, which is a
very simple, zero configuration bundling tool to reduce the amount of
time I spent debugging build issues (`node-sass`, anyone?). For now,
this is the perfect addition, though we may decide we need something
more complex in the future.

I did make a few changes to source code, namely:

- `components/loaders/loaders.hbs`
- `components/omnibar/omnibar.react.js`
- `docs/02-brand-guidelines.md`

but these were either (a) quality of life improvements (importing React
instead of duplicating it) and (b) changes to make the build system work
with paths pointing to the new `@stjudecloud/assets` package.