chore: moves bootstrap theme to the `themes/bootstrap` folder.
This commit moves all of the bootstrap related theme business to the
`themes/bootstrap` folder. This folder was initialized as a package that
is closest to the original `package.json` file (which will be deleted in
a future commit).

Notably, the package name was updated from `@stjudecloud/theme` to
`@stjudecloud/theme-bootstrap` in anticipation a `@stjudecloud/theme-tailwind`
package which will be arriving later.

Other various quality of life improvements include:

- Updating the build system to Parcel v2, for the same reasons as listed
  in the previous commit.
- Creating a `font-face` Sass mixin and using that to initialize all
  fonts in the theme.
- Updated fonts to the appropriate weights (300 should be 'light', 400
  should be 'regular', 500 should be 'medium' and was missing
  altogether, 700 should be 'bold').
- Links the `.npmrc` file to the top-level `.npmrc` to ensure we follow
  the same prerelease naming conventions.